Remove navl library and call existing untangle-classd daemon
I removed the C code that was linking and directly calling to the
Vineyard/Procera/Sandvine library, and updated the classify
plugin to use the untangle-classd daemon just like NGFW.
Still lots to do. Things will break in the classd daemon isn't
running, and I need to work on the reconnect, parsing, error
detection, etc. but it works.