Improve certificate manager
The certificate is now attached to the session object so other
apps have easy access. Started work on extracting the subject and
issuer from the cert for putting in conndict but it needs a lot
more work. Dealing with the pkix names is such a pain! I removed
the Msg prefix from all the vars in TrafficMessage because it was
not needed, and resulted in a lot of extra and cumbersome typing.