Add log level checking to LogMessage
I added a simple map for setting the log level for each of
the applications. It's static right now because I was going for
quick and simple while I work on other things. The configuration will
eventually be stored in settings. I also fixed the update_rules
script. The perr portid in /proc/net/netfilter/nfnetlink_queue
may not always match the pid of the owning process, which would
cause the rules to not be inserted. I modified the script to look
at the queue copy mode and size to determine if packetd or something
else owns the queue.