Store LogMessage configuration in a file
Added logic to store the LogMessage application level configuration
in a file. If the file doesn't exist, it is created using INFO as
the default for all known applications. I tried using gob, json, and
xml formats, but ultimately decided to use a simple appname=LEVEL
format text file to make it easy to view and edit, similar to the
log4j.xml we us in the Java world. Plus it gave me an opportunity
to work more with files, strings, maps, ranges, etc. in golang.