Change NfqueueFunc to return type NfqueueResult
This makes go enforce that a result is return no matter what, and leaves no opporutity for the plugin to forget to write to the channel. We now use a local anonymous function to write to the channel. This allows the plugin developer to just do excessive error checking and just return with no risk of deadlocking the whole daemon by forgetting to write to the channel.