Fix sni parsing and packet memory allocation
I moved the SNI extraction into a dedicated plugin because it needs
to release the session later than the certcache plugin. It's working
but I need to add a few more "is this a valid ClientHello" packet
checks. I also changed how we allocate memory for the gopacket.
Instead of doing the memory allocation in the C handler, I changed the
dispatcher to make a copy of the packet using GoBytes before returning.
This will let Go handle freeing the memory and it gives us a buffer
that has Go slice bounds checking.