Improve SNI extraction
The SNI extractor is much cleaner with better comments. I also
added a bool to the return list so we know when to release the
session in cases where we get a ClientHello that does not include
the SNI extension. I also change the reporter Event Log level
from Info to Debug.