Remove dependency on systemctl
I removed the call to systemctl for starting and stopping the classd
daemon and now directly launch the daemon during startup and kill
it during shutdown. May need more work for error detection and such.
I also added some variables to the Session table that we'll use
to release sessions after so many bytes or packets.