Improve classify plugin
The classify plugin now saves each of the details as a session
attachment. The attachments are used to determine when any of the
info has has been updated by the daemon, and we only push the
data to the conntrack dictionary when it has actually changed.
I also added logic to log an event and release the session when
the classification state reaches fully classified, or when we
have inspected 10 packets or 256k of data, whichever happens first.