Add proper IPv6 support to the conntrack functions
I removed the in_addr and in6_addr garbage from the conntrack_info
struct and went with four address fields defined as char[16] that
can hold either an IPv4 or IPv6 address. This lets us use a
Go pointer to copy the raw address direclty from the C structure
to the IP types that we pass to the conntrack subscribers. I also
pass the family to the go_conntrack_callback so we know which
family of addresses to extract. Also fixed the nf_queue_callback to
handle IPv4 or IPv6 packets.