dict: Add useful debug for entries with empty strings
Before if you tried to add a kernel dict entry
with an empty string, you got a very unhelpful
kernel printk.  Now, you'll get a message
in the packetd log telling you what field
you tried to set to an empty string, and in
which table, and with which key.