Rework certificate cache logic
I reworked the certificate cache. We only hold localMutex long enough
to make the fetch or load decision. The fetching thread will take
care of getting the certificate from the server and storing it in
the map. Other threads will return without releasing the session
so they can keep checking for the certificate to become available.
In all cases, once we have the cert, we put the details in the
dictionary and release. I also added more value length checking
to make sure we're not writing empty strings to the dict.