packetd_rules: Flush session dictionaries for new connections
We have observed occasions when a conntrack id is quickly reused, causing
stale dictionary data to be associated with a new session.  In some cases,
a stale dictionary might include "bypass_packetd = true" which means that
the new connection's traffic will not be sent to packetd at all.  This is
bad.  The new dict 'flush' command will flush the dictionary associated
with the specified table and key.  In this particular case, when a
connection is new we will flush the session dictionary for that conntrack
id before it gets queued to packetd.

MFW-95 #code-complete