Add support for playback speed multiplier
Implemented a new playspeed argument that allows setting the
traffic playback speed multiplier. The default is 1X and can be
set to any X value although at some point the multiplier will
result in a delay period shorter than the system timer resolution.
You can also pass a value of zero which will cause playback as
fast as the data can be read from the file.