Fixes for traffic capture and playback
Fixed playback memory allocation to work the same as with live traffic
where we allocate a buffer for each packet and let the go handler
free it when all processing is finished. This solved several related
problems and playback traffic now appears to be processed exactly the
same as when the it was captured. Still need to add a switch to turn
off live traffic processing during capture playback.