Use tuple instead of ctid for sessionTable index
Using the conntrack ID as the sessionTable index is still causing
grief, so we're going back to the tried and true method of creating
an index based on the tuple. The session entry gets created on the
first nfqueue packet using the client side tuple. We add another
session table index using the server side tuple when processing the
conntrack create. I think it's mostly working, but I still need to
figure out why sometimes the conntrack NEW event can't find an
existing entry in the session table. I also plan to do some work
on the Tuple String and StringReverse functions to make them as
fast and efficient as possible. My research so far says using
append() to create a byte[] array that is then returned as a string
will be the fastest option in the golang version we have.