Improve Tuple to String functions
Most everything I've read about string concatenation seems to say when
combining a small number of strings, using native += is just fine.
But since we'll be calling our Tuple-2-String functions a LOT, it
probably makes sense for us to go for as much performance as
possible. I found this page that compares three different methods.
In most cases, using bytes.Buffer seems to be the best choice:

 http://herman.asia/efficient-string-concatenation-in-go

The other advantage is that it makes the code look cleaner and
easier to read.