Add DNS name hints to all sessions
For all new sessions, the dns plugin will look for the server and
client IP addresses in our local cache, and if found we'll add
corresponding server_name_hint and client_name_hint strings
to the dictionary. I also updated the cache table cleanup logic to
expire entries based on the TTL we get from the resource record
instead of a static timeout value.