Add certificate sniffer plugin
I added a new plugin that watches for TLS traffic on TCP ports other
than 443. When detected we attempt to identify and extract the server
certificate and attach it to the session. This works most of the time
but it will sometimes fail if the server TLS negotiation messages are
spread across multiple packets that we receive in the wrong order.

I also added a CtoS boolean to NfqueueMessage, and I changed the name
of the Tuple variable to MsgTuple since the old name collided with the
type name. This also makes it easier when grepping for code that uses
the variable.

MFW-99 #code-completed