Rework certificate cache
I moved the certificate cache table to a service, and made a shared
function to attach the certificate to a session and put the useful
cert fields in the dictionary. I put the certificate prefetch code
in the certfetch plugin, and moved the tls certificate sniffing
logic to the certsniff plugin. Certificates are now cached using
the server address:port, and both plugins will use the cert from
the cache when found, or put the cert into the cache after a
successfull fetch or sniff operation.