Improve the certsniff plugin
Added very rudimentary packet re-assembly based on TCP sequence
numbers, so we'll now properly detect certificates that are larger
than a single packet, even if the packets come out of order.
The logic is simple but should almost always work. The plugin now
adds certs we passively detect to the certcache so they will be
loaded and used on subsequent connections to the same server:port.