Fix invalid memory access and other stuff
I fixed a bug in certsniff that could lead to reading past the end
of a buffer. I also changed a couple maps to store pointers instead
of objects, and added an explicit delete in cases where we
replace an existing object in a map with a new object. Added command
line options to enable golang cpu and mem profiling