Add traffic bypass flag
I added a traffic bypass flag that prevents the nfqueue, conntrack,
and netlogger C handlers from passing events to the main go callback
functions, effectively allowing us to ignore live traffic. Mainly
useful for testing and debugging, the flag can be set at startup
with a command line argument, and it can be toggled on and off
with commands to restd.