Change to use ctid instead of tuple for session map index
Reworked session table to once again use ctid instead of tuple for
the key. I removed the ICMP layers since it's not needed now, and I
simplified the findSessionEntry function to return the object or
nil which is possible now that we're returning a pointer instead
of an object. I also cleaned up setting of the ClientToServer boolean.