Traffic replay tests
I started work on the traffic replay tests. The initial test plays
a capture from a web server in Japan and checks that the geoip
sets the correct country in the dictionary. I added a new restd command
to allow the tests to query the status of the replay so they can wait
until it has finished. I need to improve the dict parsing logic and
figure out how to clear the dict before starting. Also need to figure
out where to store the playback files we'll be using.
This is a work in progress!