Improve system table flush
I reworked the table flush to simply create new session and conntrack
tables rather than walk through walk through and delete every entry
in the existing tables. It's cleaner, more efficient, and seems to be
the golang preferred approach to this kind of thing. I also moved
some of the subscription management vars and functions out of the
handler functions and into the dispatch to keep things better
organized.