Improve capture playback feature
The playback feature has been enhanced to keep a list of nfqueue and
conntrack objects that are created, and to clean them up when
playback is finished. This replaces the old method of simply wiping
the system tables. This required shuffling some things around to deal
with circular dependencies, and passing a playback boolean to the
low level go callback handlers so the housekeeping and cleanup can be
done there instead of in the C code. I also set the highest 4 bits of
every ctid during playback to reduce the chance we'll conflict with
live values on a system.