Improve packet capture and playback
I fixed the warehouse capture and playback functions to deal with the
inter-packet sleep times in nanoseconds rather than microseconds. With
less precision, the code was often not adding any delay at all, so
classification results could be random, dependent on how the processing
threads happened to be scheduled. I also tweaked the playback speed to
be a percentage, allowing playback to be faster or slower which would
be useful in detecting similar issues in the future. The default is
100 and 0 can still be used to spew the packets as fast as possible.