Add more playback tests
The playback test now checks the dictionary for values set by all
of the current (cert, clasify, dns, geoip, sni) plugins. I created
a new capture file that includes the DNS lookups for the two sessions
used for the test. I also added support for PACKETD_HOME env variable
so packetd can find the packetd_rules script no matter where the
binary is located. I added a FIXME to the warehouse script since we're
currently doing 64 bit math on a 96 bit timespec struct. In reality
it shouldn't be an issue because we're only applying a percentage
calculation to the inter-packet pause value which should always be
small enough to not cause integer overflow.