Add reverse DNS plugin
Created a new plugin to lookup and attach the reverse DNS name of the
client and server address for new sessions. A local cache file is used
for efficiency and to allow negative caching if there is no reverse.
Unrelated, I also added a panic if we detect nfqueue subscriptions
that duplicate the name of an existing subscription.

MFW-144 #code-completed