More work on capture playback
I fixed capture playback so passing zero as the playback speed
doesn't cause an exception, and now pulls packets out of the file
without pause as intended. I also updated the playback test to do
the functional test for normal, slow, and fast playback, and I set
the rest of the tests to use the zero/unlimited playback speed.