Increase nfqueue buffer and improve playback test
I changed our code to call nfnl_rcvbufsiz instead of getting the
socket descriptor and calling setsockopt since that seems to be the
approved way of doing it. I also increased our buffer from one to
four meg. I added more checks to the half and double speed playback
tests to make sure it does indeed playback the file in half the
time and twice the time of the normal playback. Approximately.