Improve playback tests
I reworked the playback tests to handle multiple files. We now pass
the file name and hash to the playback_file function, and it will
do the MD5 check and download as required when it is called, thus
eliminating the standalone download test. I then added a new playback
file to test the dns hint plugin. The playback includes a DNS query
response with 8.8.8.8 for www.yahoo.com. We then do a wget to one of
the IP addresses returned by the query, which should give us a hint
we can check for that session.