Improve the classd daemon management
I reworked the way we start and connect to the classify daemon. The
daemonManager goroutine will periodically check to see that the
daemon is running and our socket is connected. If the daemon stops it
will be restarted. If we lose our connection it will be reopened.
I also fixed the dns plugin to properly use the shutdown channel.