Fix classd session cleanup
I added a call to daemonRemove just before we release the session so
the daemon can cleanup stuff it allocated for the session. I also
added the flag to disable the timestamp in the classd log output
to cleanup our log messages a bit. I also fixed the mess.Session
checks so we're not checking for nil twice.