Fix table cleanup functions
I was trying to be efficient during table cleanup by iterating through
the table and letting the delete function handle the map mutex Lock
and Unlock. Turns out this is a bad idea as it can lead to a concurrent
map iteration and write panic. I verified all cleanup functions and
fixed several to do the right thing.