Update to use the classd -mfw flag
I modified the classify plugin to pass the -mfw flag to the classd
daemon. In this mode, traffic is passed to the classification engine
directly by the network handler, so we get the correct result on
the first packet. We no longer need to send session CREATE and REMOVE
commands, so those have been removed, and I updated the PACKET command
to include the appropriate IP4 or IP6 flag in the message header.