Fix to ignore empty classification results
When NAVL is able to classify a packet, it should always set confidence
to 50 (based on layer 3/4 info) or 100 (based on deep flow analysis).
If we see a zero in confidence, it's a good indicator that NAVL
didn't give us classification details for some reason. In these cases
we want to ignore the default values we use when creating a new
session object so we don't replace good classification details with
garbage.