Changes for interface and interface_type
Moved the extraction of the client and server interface ID and type
out of the reporter plugin and into the low level handlers which
put them in the Session object for everyone to use. Also changed the
database schema renaming client_interface to client_interace_id and
same for server to be more descriptive.