Improve handling of classification results
Instead of checking for and ignoring classification confidence zero
we now compare the reply confidence with any attached to the session.
This will ensure we don't use the /IP defaults if we have any
info at all, and it will ensure we don't apply a lower confidence
reply on top of a higher confidence already attached.