Improve attachment handling
We weren't holding the attachments Mutex while evaluating the existing
confidence and conditionally applying the updates which I think was
causing threads to sometimes step on each other while updating the
classification details. I added a new Session.LockAttachments()
function to dispatch that will lock the attachments Mutex and return
the attachments map for direct manipulation. We hold the lock while
checking and applying changes, and call Session.UnlockAttachments()
to release the lock when we are finished. It should also be more
efficient, since the first update was previously calling Lock/Unlock
six times for each of the details we initially set.