Add flag to use external classd daemon
I added a boolean we can manually set to disable starting and stopping
the classd daemon for use when we want to valgrind classd running with
packetd. I added a write to the daemonChannel after Dial() errors to
trigger the reconnect attempt rather than waiting for the next nfqueue
message. I also increased the size of the channel buffer to make sure
we never block when waking up the daemon checker, and made the check
logic sleep a bit longer after reading the channel.