Work on the stats plugin
Fixed a nil pointer bug when getting the session attachment. For the
time being we're only using the first xmit/recv latency measurement
but I still think we may need to revisit this. I often see a large
difference between first and subsequent connections to the same
resource. Maybe we deal with it in the moving average calcuations?
Speaking of which, I also added a MovingAverage type. Right now
it's just a brute force implementation that stores the values in
a list and calculates the average on demand. I also included the
ability to get the average for the most recent nnn samples so we
can get the last 10, 100, 1000, etc. average all from the same object.