Fix daemon restart
On daemon process exit we weren't setting daemonProcess to nil so the
manager goroutine never realized that a restart was needed. I added
a call to clear the pointer after the process exits so restart will be
triggered as long as the shutdown flag isn't set. I also changed some
of the function names to better reflect what each function actually
does to make it easier to follow the logic. I also removed two
FIXME items in other places for things that I have confirmed are
fine just the way they are.