Rework process and socket management to eliminate data races
The daemon process and socket each have their own goroutine that is
responsible for monitoring and restarting or reconnecting. Each waits
on a channel for messages that signal what action to take. I played a
lot with using a channel to send the classify requests to the daemon
and another channel to get the result, but I saw issues with latency
under very heavy load, so I stuck with our existing approach of using
a sync.Mutex to protect the net.Conn socket object and sending the
raw packets directly from the nfqueue handler. In the daemon process
manager I used a sync.WaitGroup to know when the stdin and stdout
readers are finished so we don't touch the cmd.Exec object while those
goroutines are still active, which also made the overall flow cleaner
and more logical. Finally, I moved the process and socket managers into
their own source files to make the code easier to read and maintain,
and all detected data races have now been resolved.