Fix data races in conntrack
I added a RWMutex to the Conntrack structure. Anything that uses
an active entry must ensure it wraps access in calls to RLock/RUnlock
or Lock/Unlock as appropriate. I also fixed the GetConntrackTable()
function so it holds the conntrackTableMutex while copying the table
to prevent concurrent map access. Finally, I did some cleanup in
StopCallbacks so if we do hang during shutdown, we can tell which
of the tasks we wait on is actually hung.