Fix missing stats_holder issue
Instead of creating our own stats_holder attachment with the time when
we see the first client packet, I changed the code to use the session
creation time for calculating latency, since it should effectively
be the same as what we were attaching. I also adjusted all of the
plugin priorities so stats gets called first so we get the most
accurate measurement possible. To make it easier to manage the
priorities in the future, I removed hard coded priority values in all
of the InsertNfqueueSubscription calls, and replaced them with
dispatch consts which we can see and adjust in a single place.