Add package to manage simple named counters
I added a package to keep and manage a simple list of named counters
and the ability to view them by going to ipaddr:/debug

The catalyst was the flood of log messages that are generated when
testing under very heavy load. There is currently only one counter
in reports.LogEvent but I anticipate adding more in areas where we
log critical issues for better visibility.