Improve support for multiple ping targets
I reworked the way we handle ICMP for active latency ping sampling.
At the configured interval, we send all of the ping packets, and
let goroutines listening on each of the interfaces handle
receiving the replies, preforming the latency calculation, and
recording the value for each interface. Timeouts from the previous
sample period are detected and logged just before the new batch of
ping probes is sent out.