Add support for HUP signal
I added a handler for the HUP signal which is currently of interest
only to the stats plugin. On HUP stats will close the sockets
we use for active ping tests, reload the network interface details,
and then re-open the sockets using the new interface configuration.
Rather than a bloated subscriber model, I used the simplified
model of a known function name. Much like all plugins must have
a PluginStartup() and PluginShutdown() function, they can also
have an optional PluginSignal() function. If a handler is added
to a plugin, a call to the function should be added to the
signalPlugins() function.