Add XL and XU country codes
I reworked the geoip plugin to always assign a country code to the
client and server sides of a session. We start with XU for unknown
and then assign XL for addresses in RFC1918 or other local blocks
suck as loopback. Everything else is checked against the GeoIP
database and we use any country code it gives back.

MFW-351 #code-completed