Improve classification plugin
When creating the string from the classd daemon response stored in
daemonBuffer, the code wasn't using the socket read length to limit
the data pulled from the buffer, so the string was always 1024 bytes
in size. This could cause random cruft past the end of the received
message to match a tag in our reply parser, possibly overwriting one
of the values we extract. Fixed by limiting the string to the actual
length of data read from the socket. Classification is also improved
by always using the client side tuple in the raw packet we pass to the
NAVL library. This solves the problem of pre-nat and post-nat traffic
being evaluated as two completely different sessions. It should give
us better results and eliminates some of the previous hacks we used to
determine the "best" result from the two. The basic mechanism to
detect matched vs. inferred differences is also in place but I still
need to add logic to collect the details and push them to the cloud.

@changelog