Change reports -disable-cloud to global -no-cloud
When developing and debugging, it is often desirable to prevent
packetd from hitting the cloud. Now that we have several things that
connect to the cloud, I removed the private flag we had in the reports
service and implemented a global FlagNoCloud boolean in kernel.
It is now checked by everything that connects to or requires the
cloud (except cmd auth) allowing everything to be controlled
with a single flag.