Fix queue range passed to packetd_rules script
We had a one vs zero based math error in our rules script that was
causing the fanout statement to spread the load across one more
queue than we have actual listening threads. The nft docs suggest
that fanout uses the CPUID to balance the load, so in practice it
should never have pushed traffic to the unserviced queue, but better
safe than sorry. Plus, it's possible to use a queue range without
fanout, in which case it would (I assume) push traffic by cycling
through the defined queues in order. If we ever decide to try that,
the math is now correct. I also did some general cleanup in the
packetd_rules script to make it easier to read and maintain.
@changelog