Add disparity counters for event subcriptions
I added overseer disparity counters that are incremented when we call
an event subscriber, and decremented when the function returns. This
will allow us to track down plugins or services that get stuck and
never return. To make this mechanism as efficient as possible I
improved the overseer class so the map stores integer pointers and
is protected by a sync.RWMutex that only requires a full lock when
a named counter is first created. All other map operations only need
a read lock, and all of the counter values are accessed atomically.
@changelog