Fix timeout handling for slow nfqueue subscribers
I finally found and fixed a bug that has been around forever. When
handling an nfqueue subscriber timeout, the callSubscribers function
was writing the force release result to the wrong channel. This
filled that channel, which would hang the subscriber goroutine when
the subscriber finally returned and tried to write the subscriber
result to the now full channel. It would also cause callSubscribers
to hang waiting on the result channel that was underfilled when the
force release result was written to the wrong channel. So each
nfqueue subscriber timeout would result in two goroutines that would
stick around forever. I fixed the code to write the timeout result
to the correct channel, and reworked the timeoutTimer logic so the
subscriber function goroutine doesn't try to write to the inner result
channel if the timeout was triggered. I also removed the defer syntax
in ReleaseSession and now call Unlock directly, as I've recently
learned that defer has a lot of runtime overhead and should not be
used in a hot code path. Finally, I removed the deep scan log
warning since it's not accurate and creates noise. First, to implement
correctly would require another lock on the session subscriber table
to get an accurate count. Second, it's possible we'll have more than
the trigger threshold queued for processing before we finish handling
all of the unsubscription requests and push the bypass flag to dict.

@changelog