Return correct network details for bridged devices
I added a bridge map that is built from the bridge-members section
of any network.device entries. When adding the network info for
each device, we check to see if it's listed as a bridge-member for
any of the network interfaces. If so, we pull the IP, DNS and gateway
info from that network config section. Otherwise we just look for the
config for the actual named device.