More logger cleanup and optimization
logger.go - I moved functions around putting the public functions at
the top, with private functions lower. I changed some stuff that was
public to private to ensure findCallingFunction() will always called
such that the external calling function is the 5th frame on the call
stack. This is to make sure the log level lookup always works as
expected. I also tuned that function to only retrieve a single frame.
I added a source string to the LogWriter structure so we can set and
control the log output of anything that uses the stream writer
mechanism.

restd.go - I added the source to the LogWriter we create to capture
gin system output. I also changed the other places where we capture
output from gin to call LogMessageSource so we can now control the
gin output directly rather than it being treated as restd output.

nfqueue.go - I removed the timeMap and timeMapLock stuff that was
calculating the execution time of each subscriber. This is left over
from early debugging of the dispatcher, and was expensive logic
to leave unused in the most active code path in the daemon.

classify.go - corrected use of now private IsLogEnabled function

settings.go - added source in call to logger.NewLogWriter

packetd.go - fixed misspelling