Remove subscriber disparity counters
Now that the nfqueue dispatch loop is fixed, the subscriber enter/exit
counters I added to help track down the problem are redundant, and add
noise to the /api/debug page. The nfqueue subscriber call loop
already logs any timeout using an OC verb which inclues the subscriber
name in the counter, and the conntrack and netlogger dispatchers
have similar timeout detection and logging as well.