sysupgradeHandler: Return better errors
Be a bit more helpful with the errors we return to the UI.  If
we fail to upload/create/copy the flash file, return an upload
error.  If we test the flash file and can determine that this image
doesn't support this device, return an incorrect upgrade image error.
And finally, if the actual sysupgrade fails, just return upgrade failed.