Turn down some diagnostic logging
Now that I'm narrowing down the 'database is locked' issue to the trim
and vacuum code, I'm changing some of the logging I turned up back
down to the debug level. I also changed the dictionary cleanup to
the debug level as well. My goal is to make sure when a trim and vac
finally happens, we don't want extra noise pushing the messages and
timestamps out of the buffer, since we need that info to figure out
which part of the cleanup is holding the DB locked for so long.