Disable database shared cache mode
I enabled shared cache mode based on earlier research into using
sqlite in a multi-threaded enviornment. Unfortunately we're still
seeing issues on live sites with 'database table is locked' errors
that seem to only happen up when a UI connection is also active.
Further research led me to a different discussion that suggested why
shared cache mode can actually cause more problems.

https://github.com/mattn/go-sqlite3/issues/632

From this discussion, it seems disabling shared cache mode but keeping
the busy timeout we now set may be the best solution.
Ultimately, it seems multi-threaded access is an ongoing issue for
many people, and the best solution may be to rework the model we
use for handling the DB queries we do for the user interface.