Add logic to retry lookups that encounter comm failure
If there is a failure sending a prediction request to the cloud
server, we now push the request back in the request channel to allow
retry. I added a hitCount field to the request structure so we can
keep track of the number of retry attempts, and eventually timeout
and return the Unknown response once the limit is reached.