Add session keep-alive
The cloud server seems to disconnect our persistent connections
after a period of time with no traffic. I added a timeout case to
the main request channel handler so we'll generate a dummy lookup
request on our persistent connections when we go for too long without
servicing an actual request. In testing, the server timeout seems to
be 60 seconds so I used 45 seconds as the timeout value before sending
the keep-alive lookup.