Fix source for interface status Connected boolean
I changed the /api/status/interfaces call so the Connected boolean
will now be set using the "carrier" field from the ubus result
instead of the "up" field. This should give us the same behavior
as NGFW where the interface status icon shows green only if the
interface is up AND is connected to something.