Modify predict service to use persistent transport and client
The existing logic is right on the money. We just needed to move
the creation of the http transport and client objects into the
startup function so they can persist and be shared, and tweak the
timeout and idle connection limits to complement our use case.
According to the golang net/http docs, Clients and Transports are
safe for concurrent use by multiple goroutines and for efficiency
should only be created once and re-used.