Add Keep-Alive header and disable ExpectContinue
I added Connection: Keep-Alive to the request header. Shouldn't be
required as HTTP/1.1 connections are supposed to assume keep-alive
by default but it seemes to improve underlying socket persistence,
and certainly can't hurt. I also set ExpectContinueTimeout to zero
in our transport. This is a mechanism that is designed to avoid
sending a large request body to a server that may not be able to
process it. It sends the header and waits for the server to say
continue before sending the body. It's not really needed in our
case so disabling reduces client/server chatter and improves
overall throughput.