Add logic to handle multiple simultaneous requests
We typically see simultaneously connections to the same server.
Previously, until one of the prediction requests completed and put
the result in the cache, every subsequent request for the same server
would fire off another request to the cloud. I reworked the code to
use the same solution we used for the same problem in certfetch.
I added a WaitGroup to the cache holder which allows the first thread
to send the request to the cloud, while allowing subsequent threads
needing the same result to wait for the outstanding request to
complete instead of submitting a duplicate request. I also reworked
the cache timeout a bit. I made the normal timeout static to ensure
we go back to the cloud for fresh results periodically. I also added
a much shorter timeout we use when storing Unknown in response to a
network problem so transient connectivity issues won't leave long
lived Unknown entries hanging around for too long. Also fixed the
GET reply logic to make sure we always read the full http response
body per the golang docs.