Improve handling of prediction cache times
Based on code review feedback, I added another cache time. We cache
good cloud results for 1 day. If we talk to the cloud but we get
something other than http.StatusOK we cache the unknown result
for 1 hour. If we encounter an error connecting to the cloud or
reading the response object, we cache the unknown result for 1 minute.
I also simplified the logic so the main traffic handler just uses the
classification info and cache time returned from the lookup function
since all the logic is already built into that code path.