Improve the database lookup functions
While looking at the source for the geoip2 package I noticed that
calling geoip2.City causes the code to attempt parsing a bunch
of extra fields that won't exist in the smaller country database
we are using. To make our lookups more efficient and match the
database we're using, I changed our code to use geoip2.Country
for lookups.