Fix obscure invalid memory access during startup
When the nfqueue code was enhanced to service multiple queues, one
of the array references did not get updated, causing us to check
nfqqh for zero instead of nfqqh[index]. This broke the return value
check which ultimately caused an exception when trying to access
a queue handle that wasn't successfully opened, like would happen when
starting a second instance of packetd when one is already running.
This wouldn't happen in normal circumstances, but I just spent several
hours thinking there was some other problem in the new development
environment I had just setup. The real issue was this bug prevented
packetd from doing the expected graceful shutdown when it detects
this situation, because it wasn't properly detecting it.

@changelog